Kernel - 95178dae7e60733ed3848d66ad813c97f09995c9, Bootstrap - 123e82051c4ebe81ee4ef234f0d9d91e1b14e77e, PortableStructsGen - 162a2dbd00def4f2b1a63281f7699fd398f9edda, LibLinux - update in progress

+++ headers/_/_generic_bootstrap.h
@@ -95,9 +95,8 @@ typedef struct
 	void *(*file_open_readwrite)(const char *path);
 	void *(*file_open_append)(const char *path);
 
-	long long (*file_read)(void *file, unsigned long long offset, unsigned char *data, size_t size);
-	long long(*file_write)(void *file, unsigned long long offset, unsigned char *data, size_t size);
-
+	int (*file_read)(void *file, unsigned long long offset, unsigned char *data, size_t size);
+	int (*file_write)(void *file, unsigned long long offset, unsigned char *data, size_t size);
 	long (*file_mode)(const char * file);
 	long (*file_length)(const char * file);
